Increase queueing capacity (#5)
* Refactor tests

* Increase queueing capacity

Originally only one batch could be queued up at a time.

If the available processing concurrency is e.g. 10 batches of 10 items,
then the worst case is that all batches finish at the same time and we
only have 10 new items to process, using only 10% of the available
capacity.

If we can queue up the same number of items as we can process, then we
would be able to use 100% of the processing capacity in this case.

To that end, we now queue multiple batches ready for processing, instead
of just one.

* Improve instant comparison

* Put plaster over test flakiness

https://github.com/tokio-rs/tracing/issues/2874

Looks like a bug in tracing is making the tests flaky.

* Reenable win-gnu