Implement resource acquisition
To decrease batch concurrency when we get large spikes of jobs with an
empty database connection pool.

When using the Immediate strategy, this would immediately try to
maximise the number of database connection, which can put strain on the
system.

Here, we try to acquire any resources (e.g. database connections) ahead
of processing, and let new items get added to the batch while waiting.